+++ src/components/app-menu-container/__tests__/menu.test.js
@@ -5,28 +5,29 @@ import { HashRouter as Router } from 'react-router-dom'
 import ListItem from '@material-ui/core/ListItem'
 
 import AppMenuContent from '../menu'
+import PAGES from '../../../pages/pages-config'
 
 describe('Component/AppMenuContent', () => {
   test('should render without throwing an error', () => {
-    expect(shallow(<AppMenuContent />)).toMatchSnapshot()
+    expect(shallow(<AppMenuContent pages={PAGES} />)).toMatchSnapshot()
   })
 
   test('should render the proper number of ListItems', () => {
     const wrapper = mount(
       <Router>
-        <AppMenuContent />
+        <AppMenuContent pages={PAGES} />
       </Router>,
     )
-    expect(wrapper.find(ListItem).length).toBe(4)
+    expect(wrapper.find(ListItem).length).toBe(3)
   })
 
   test('should render the proper order of ListItems', () => {
     const wrapper = mount(
       <Router>
-        <AppMenuContent />
+        <AppMenuContent pages={PAGES} />
       </Router>,
     )
-    expect(wrapper.find(ListItem).length).toBe(4)
+
     expect(
       wrapper
         .find(ListItem)
@@ -38,18 +39,12 @@ describe('Component/AppMenuContent', () => {
         .find(ListItem)
         .at(1)
         .prop('to'),
-    ).toBe('/config')
+    ).toBe('/github')
     expect(
       wrapper
         .find(ListItem)
         .at(2)
         .prop('to'),
     ).toBe('/privacy')
-    expect(
-      wrapper
-        .find(ListItem)
-        .at(3)
-        .prop('to'),
-    ).toBe('/github')
   })
 })