+++ src/constants/index.js
@@ -1,6 +1,6 @@
 export const GITHUB_URL = 'https://github.com/natterstefan/react-trello-multiboard/'
 
-export const breakpoints = {
+export const BREAKPOINTS = {
   small: '@media (max-width: 768px)',
   medium: '@media (max-width: 1440px)',
   large: '@media (min-width: 1441px)',