+++ src/pages/__tests__/page-config.test.js
@@ -4,24 +4,14 @@ import { HashRouter as Router } from 'react-router-dom'
 import SyntaxHighlighter from 'react-syntax-highlighter'
 
 import ConfigPage from '../page-config'
-import Layout from '../../layout'
 
 describe('Page/Config', () => {
   test('should render without throwing an error', () => {
     const wrapper = shallow(<ConfigPage />)
-    expect(wrapper).toHaveLength(1)
+    expect(wrapper).toMatchSnapshot()
   })
 
-  test('should render a Layout', () => {
-    const wrapper = mount(
-      <Router>
-        <ConfigPage />
-      </Router>,
-    )
-    expect(wrapper.find(Layout).length).toBe(1)
-  })
-
-  test('should render with SyntaxHighlighter as a child of Layout', () => {
+  test('should render with SyntaxHighlighter as a child of the page', () => {
     const wrapper = mount(
       <Router>
         <ConfigPage />