+++ src/pages/__tests__/page-privacy.test.js
@@ -1,8 +1,7 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 
-import PrivacyPage, { PrivacyContent } from '../page-privacy'
-import Layout from '../../layout'
+import PrivacyPage from '../page-privacy'
 
 describe('Page/PrivacyPage', () => {
   test('should render without throwing an error', () => {
@@ -10,13 +9,9 @@ describe('Page/PrivacyPage', () => {
     expect(wrapper).toHaveLength(1)
   })
 
-  test('should render a Layout', () => {
-    const wrapper = shallow(<PrivacyPage />)
-    expect(wrapper.find(Layout).length).toBe(1)
-  })
-
   test('should render with PrivacyContent as a child of Layout', () => {
     const wrapper = shallow(<PrivacyPage />)
-    expect(wrapper.dive().find(PrivacyContent).length).toBe(1)
+    expect(wrapper.find('#privacy').length).toBe(1)
+    expect(wrapper.find('#privacy').html()).toContain('Hello Privacy.md!')
   })
 })