+++ src/pages/page-app.js
@@ -1,14 +1,9 @@
 import React from 'react'
 
 // Components
-import Layout from '../layout'
 import MainApp from '../components/main-app'
 
-const AppPage = props => (
-  <Layout>
-    <MainApp {...props} />
-  </Layout>
-)
+const AppPage = props => <MainApp {...props} />
 AppPage.displayName = 'AppPage'
 
 export default AppPage